c file

+++ 0x0C-more_malloc_free/101-mul.c
@@ -9,7 +9,7 @@
 int main(int argc, char *argv[])
 {
 	int i = 0, j = 0;
-	int result;
+	unsigned long result;
 
 	if ((argc - 1) != 2)
 	{
